vertex and edge selection now visible  in GUI

+++ src/view/GUI.hpp
@@ -26,7 +26,7 @@ namespace graphvise
     private:
 
         double fps = 0.0;
-        int currentObjId = 0;
+        uint64_t currentObjId = 0;
         Buttons buttons;
         bool errorAvailable;
         std::optional<Error> currentError = std::nullopt;
